\name{endomer-package}
\alias{endomer}
\alias{endomer-package}
\title{Metapackage for ENFT, ENCFT, ENHOGAR and ENGIH}
\description{Declare compatible minimum versions and attach the ENFT, ENCFT, ENHOGAR and ENGIH interfaces. Inspect versions and prepare update reports without automatically changing installed packages.}
