\name{vkGetAdCampaigns}
\alias{vkGetAdCampaigns}
\title{Р’РѕР·РІСЂР°С‰Р°РµС‚ СЃРїРёСЃРѕРє СЂРµРєР»Р°РјРЅС‹С… РєР°РјРїР°РЅРёР№}
\description{Р¤СѓРЅРєС†РёСЏ vkGetAdCampaigns РІРѕР·РІСЂР°С‰Р°РµС‚ СЃРїРёСЃРѕРє РґРѕСЃС‚СѓРїРЅС‹ СЂРµРєР»Р°РјРЅС‹С… РєР°РјРїР°РЅРёР№ РёР· СЂРµРєР»Р°РјРЅРѕРіРѕ РєР°Р±РёРЅРµС‚Р° Р’РєРѕРЅС‚Р°РєС‚Рµ.}
\usage{
vkGetAdCampaigns(account_id = NULL, client_id = NULL, include_deleted = TRUE, campaign_ids = "null", access_token = NULL)
}
\arguments{
  \item{account_id}{РРґРµРЅС‚РёС„РёРєР°С‚РѕСЂ СЂРµРєР»Р°РјРЅРѕРіРѕ РєР°Р±РёРЅРµС‚Р°, СЃРїРёСЃРѕРє РІСЃРµС… РґРѕСЃС‚СѓРїРЅС‹С… СЂРµРєР»Р°РјРЅС‹С… РєР°Р±РёРЅРµС‚РѕРІ РјРѕР¶РЅРѕ РїРѕР»СѓС‡РёС‚СЊ СЃ РїРѕРјС‰СЊСЋ С„СѓРЅРєС†РёРё vkGetAdAccounts.}
  \item{client_id}{РРґРµРЅС‚РёС„РёРєР°С‚РѕСЂ РєР»РёРµРЅС‚Р°, Сѓ РєРѕС‚РѕСЂРѕРіРѕ Р·Р°РїСЂР°С€РёРІР°СЋС‚СЃСЏ СЂРµРєР»Р°РјРЅС‹Рµ РѕР±СЉСЏРІР»РµРЅРёСЏ. Р”РѕСЃС‚СѓРїРЅРѕ Рё РѕР±СЏР·Р°С‚РµР»СЊРЅРѕ РґР»СЏ СЂРµРєР»Р°РјРЅС‹С… Р°РіРµРЅС‚СЃС‚РІ.}
  \item{include_deleted}{Р›РѕРіРёС‡РµСЃРєРѕРµ TRUE РёР»Рё FALSE, С„Р»Р°Рі, Р·Р°РґР°СЋС‰РёР№ РЅРµРѕР±С…РѕРґРёРјРѕСЃС‚СЊ РІС‹РІРѕРґР° Р°СЂС…РёРІРЅС‹С… РѕР±СЉСЏРІР»РµРЅРёР№.}
  \item{campaign_ids}{Р§РёСЃР»РѕРІРѕР№ РІРµРєС‚РѕСЂ, С„РёР»СЊС‚СЂ РїРѕ СЂРµРєР»Р°РјРЅС‹Рј РєР°РјРїР°РЅРёСЏРј.}
  \item{status_names}{Р›РѕРіРёС‡РµСЃРєРѕРµ TRUE РёР»Рё FALSE, РѕС‚РІРµС‡Р°РµС‚ Р·Р° С„РѕСЂРјР°С‚ РІС‹РІРѕРґР° СЃС‚Р°С‚СѓСЃР° РєР°РјРїР°РЅРёРё, РµСЃР»Рё TRUE СЃС‚Р°С‚СѓСЃ Р±СѓРґРµС‚ РІРѕР·РІСЂР°С‰С‘РЅ РІ РІРёРґРµ С‚РµРєСЃС‚РѕРІРѕР№ СЃС‚СЂРѕРєРё, РµСЃР»Рё FALSE С‚Рѕ Р±СѓРґРµС‚ РІРѕР·РІСЂР°С‰С‘РЅ РЅРѕРјРµСЂ СЃС‚Р°С‚СѓСЃР°.}
  \item{access_token}{РўРѕРєРµРЅ РґРѕСЃС‚СѓРїР° Рє API, РїРѕР»СѓС‡РµРЅРЅС‹Р№ СЃ РїРѕРјРѕС‰СЊСЋ С„СѓРЅРєС†РёР№ vkAuth РёР»Рё vkGetToken}
}
\author{
Alexey Seleznev
}
\value{Р”Р°С‚Р° С„СЂРµР№Рј СЃРѕ СЃРїРёСЃРєРѕРј СЂРµРєР»Р°РјРЅС‹С… РєР°Р±РёРЅРµС‚РѕРІ Рё РёС… РїР°СЂР°РјРµС‚СЂРѕРІ СЃРѕ СЃР»РµРґСѓСЋС‰РёРјРё Р·РЅР°С‡РµРЅРёСЏРјРё:
\item{1. id}{РёРґРµРЅС‚РёС„РёРєР°С‚РѕСЂ РєР°РјРїР°РЅРёРё.}
\item{2. type}{С‚РёРї РєР°РјРїР°РЅРёРё.}
\item{3. name}{РЅР°Р·РІР°РЅРёРµ РєР°РјРїР°РЅРёРё.}
\item{4. status}{СЃС‚Р°С‚СѓСЃ РєР°РјРїР°РЅРёРё. }
\item{5. day_limit}{РґРЅРµРІРЅРѕР№ Р»РёРјРёС‚ РєР°РјРїР°РЅРёРё РІ СЂСѓР±Р»СЏС…. }
\item{6. all_limit}{РѕР±С‰РёР№ Р»РёРјРёС‚ РєР°РјРїР°РЅРёРё РІ СЂСѓР±Р»СЏС…. }
\item{7. start_time}{РІСЂРµРјСЏ Р·Р°РїСѓСЃРєР° РєР°РјРїР°РЅРёРё. }
\item{8. stop_time}{РІСЂРµРјСЏ РѕСЃС‚Р°РЅРѕРІРєРё РєР°РјРїР°РЅРёРё. }
\item{9. create_time}{РІСЂРµРјСЏ СЃРѕР·РґР°РЅРёСЏ РєР°РјРїР°РЅРёРё. }
\item{10. update_time}{РІСЂРµРјСЏ РїРѕСЃР»РµРґРЅРµРіРѕ РёР·РјРµРЅРµРЅРёСЏ РєР°РјРїР°РЅРёРё. }
}
\examples{
##РђРІС‚РѕСЂРёР·Р°С†РёСЏ РІ РІРє
my_tok <- vkAuth(app_id = 1,app_secret = "H2Pk8htyFD8024mZaPHm")

##Р—Р°РїСЂРѕСЃ СЃРїРёСЃРєР° СЂРµРєР»Р°РјРЅС‹С… РєР°РјРїР°РЅРёР№
my_vk_camp <- vkGetAdCampaigns(account_id = 1, 
                               access_token = my_tok$access_token)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
