\name{vkGetUserGroups}
\alias{vkGetUserGroups}
\title{Список сообществ Вконтакте.}
\description{Функция vkGetUserGroups Возвращает список сообществ указанного пользователя.}
\usage{
vkGetUserGroups(user_id = NULL, filter = NULL, access_token = NULL)
}
\arguments{
  \item{user_id}{идентификатор пользователя, для которого необходимо получить список друзей. Если параметр не задан, то считается, что он равен идентификатору текущего пользователя.}
  \item{filter}{список фильтров сообществ, которые необходимо вернуть, перечисленные через запятую. Доступны значения admin, editor, moder, groups, publics, events. По умолчанию возвращаются все сообщества пользователя. При указании фильтра admin будут возвращены сообщества, в которых пользователь является администратором, editor — администратором или редактором, moder — администратором, редактором или модератором.}
  \item{access_token}{Токен доступа к API, полученный с помощью функций vkAuth или vkGetToken}
}
\value{Дата фрейм со статистикой сообщества со следующими значениями:
\item{1. gid}{идентификатор сообщества.}
\item{2. name}{название сообщества.}
\item{3. screen_name}{короткий адрес, например, apiclub.}
\item{4. is_closed}{является ли сообщество закрытым. Возможные значения: 0 — открытое, 1 — закрытое, 2 — частное}
\item{5. type}{тип сообщества: group — группа, page — публичная страница, event — мероприятие.}
\item{6. description}{текст описания сообщества.}
\item{7. wiki_page}{название главной вики-страницы.}
\item{8. members_count}{количество участников.}
\item{9. start_date}{дата основания.}
\item{10. can_post}{информация о том, может ли текущий пользователь оставлять записи на стене сообщества. Возможные значения: 1 — может, 0 — не может.}
\item{11. can_see_all_posts}{информация о том, разрешено ли видеть чужие записи на стене группы. Возможные значения: 1 — может, 0 — не может.}
\item{12. activity}{строка состояния публичной страницы. У групп возвращается строковое значение, открыта ли группа или нет, а у событий дата начала.}
\item{13. status}{статус сообщества.}
\item{14. fixed_post}{идентификатор закрепленной записи.}
\item{15. verified}{информация о том, верифицировано ли сообщество. Возможные значения: 1 — может, 0 — не может.}
\item{16. site}{адрес сайта из поля «веб-сайт» в описании сообщества.}
\item{17. can_create_topic}{информация о том, может ли текущий пользователь создать новое обсуждение в группе. Возможные значения: 1 — может,  0 — не может.}
\item{18. photo}{URL главной фотографии с размером 50x50px.}
\item{19. photo_medium}{URL главной фотографии с размером 100х100px.}
\item{20. photo_big}{URL главной фотографии в максимальном размере.}
}
\author{
Alexey Seleznev
}

\examples{
##Авторизация в вк
my_tok <- vkAuth(app_id = 1, app_secret = "H2Pk8htyFD8024mZaPHm")

##Получаем список групп в которы состоит пользователь пользователя
my_vk_community <- vkGetUserGroups(user_id = 7437995,
                                   access_token = my_tok$access_token)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
